add reaper support
