+++ fflogs_rotation/job_data/potencies/melee/Reaper/7_01-Reaper.csv
